+++ test/locales/translations.test.js
@@ -111,6 +111,7 @@ const ignoredTransKeyForYourLoops = [
   "consent-hcp-terms-of-use",
   "consent-patient-privacy-policy",
   "consent-patient-terms-of-use",
+  "modal-add-medical-team-code-no-invite",
   "signup-steppers-caregiver-title",
   "signup-steppers-hcp-title",
   "signup-steppers-step1",