+++ CHANGELOG.md
@@ -18,7 +18,9 @@ It is based on Tidepool Blip 1.27.
 - YLP-1475 Implement the team details page for patients
 - YLP-1476 Remove care teams page
 - YLP-1494 Implement patient table for caregiver
+- YLP-1495 Add success/error messages on team details page
 - YLP-1496 Add possibility for a patient to add a caregiver
+- YLP-1500 Patient can join a team from header
 - YLP-1511 Display message date, status and the number of unread messages in chat widget
 ### Engineering Use
 - YLP-1256 Split Profile component into smaller subcomponents